Move menu_bar.py into RimSort.views
