+++ RimSort/views/main_window.py
@@ -15,7 +15,7 @@ from RimSort.utils.system_info import SystemInfo
 from util.watchdog import RSFileSystemEventHandler
 from view.game_configuration_panel import GameConfiguration
 from view.main_content_panel import MainContent
-from view.menu_bar import MenuBar
+from RimSort.views.menu_bar import MenuBar
 from view.settings_dialog import SettingsDialog
 from view.status_panel import Status
